requirements.txt
tensorflow-gpu==1.12.0
bert4keras
numpy
gensim
